Move pageMeta settings to drawing reducer
